requirements.txt
Flask
Flask-SocketIO
python-dotenv
Flask-Session